Evaluate properties only if necessary.
If properties were set using actual objects instead of procs or blocks,
then indirection of creating and later evaluation of procs just costs time
and adds no additional flexibility.

Thus, we can just check if a property value needs to be evaluated and either
do so or just return the final value. `eval_property` is introduced to allow
this conditional evaluation.